Изменен тип цены Добавлена миграция

+++ WebScraper.WebApi/DTO/PriceDto.cs
@@ -11,8 +11,8 @@ namespace WebScraper.WebApi.DTO
     {
         [Key]
         public int Id { get; set; }
-        public int Price { get; set; }
-        public int? DicountPrice { get; set; }
+        public decimal Price { get; set; }
+        public decimal? DicountPrice { get; set; }
         public double? DiscountPercentage { get; set; }
         public DateTime Date { get; set; }
